move readSBOM to the gokrazy package
